+++ RELEASE-NOTES.md
@@ -10,7 +10,7 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 ## [Unreleased] - Release date yyyy-mm-dd
 
 ### Fixed
-- Guard HIP compiler flag ``--rocm-path=/path/to/rocm`` against Crayftn compiler.
+- Guard HIP compiler flag ``--rocm-path=/path/to/rocm`` against Crayftn compiler earlier than 15.0.0.
 
 ## [Version 0.5.2] - Release date 2022-10-05