+++ RELEASE-NOTES.md
@@ -9,6 +9,9 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 
 ## [Unreleased] - Release date yyyy-mm-dd
 
+### Changed
+ - Updated Googletest to [12a5852e451baabc79c63a86c634912c563d57bc](https://github.com/google/googletest/commit/12a5852e451baabc79c63a86c634912c563d57bc), at date 04/13/2023.
+
 ### Added
 - Added `blt_print_variables` macro to print variables in current scope, with regex filtering on variable names and values