+++ RELEASE-NOTES.md
@@ -13,6 +13,7 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 - Updated Googletest to main from 04/13/2023.
   Commit: [12a5852e451baabc79c63a86c634912c563d57bc](https://github.com/google/googletest/commit/12a5852e451baabc79c63a86c634912c563d57bc).
   Note: this version of Googletest requires C++14, and PGI is not supported. If you are using PGI, set ENABLE_GTEST OFF.
+- Updated GoogleBenchmark to 1.8
 
 ### Added
 - Added `blt_print_variables` macro to print variables in current scope, with regex filtering on variable names and values